Change unsigned to size_t

+++ src/xmipp/libraries/reconstruction_cuda/cuda_forward_art_zernike3d.h
@@ -58,7 +58,7 @@ class Program {
 		size_t idxY0, idxZ0;
 		PrecisionType iRmaxF;
 		PrecisionType *cudaClnm, *cudaR;
-		unsigned clnm_size;
+		size_t clnm_size;
 	};
 
    public:
